complete validation and reusability
