+++ package.json
@@ -7,7 +7,7 @@
     "autoprefixer": "^9.8.6",
     "cra-template": "1.1.0",
     "postcss-cli": "^8.1.0",
-    "prettier": "^2.0.5",
+    "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
@@ -40,5 +40,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "prettier": "^2.0.5"
   }
 }
