+++ src/Dashboard.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Dashboard() {
+  return <div>Welcome to Softcom</div>;
+}
